Move out all the mappings
