+++ resources/assets/lib/notification-widget/item-one.tsx
@@ -18,29 +18,13 @@
 
 import * as _ from 'lodash';
 import { observer } from 'mobx-react';
-import LegacyPmNotification from 'models/legacy-pm-notification';
+import { nameToIcons } from 'notification-maps/icons';
+import { messageOne } from 'notification-maps/message';
+import { urlOne } from 'notification-maps/url';
 import * as React from 'react';
 import { Spinner } from 'spinner';
 import ItemProps from './item-props';
-import { one as url } from './url';
 import { withMarkRead, WithMarkReadProps } from './with-mark-read';
-
-interface IconsMap {
-  [key: string]: string[];
-}
-
-const ITEM_NAME_ICONS: IconsMap = {
-  beatmapset_discussion_lock: ['fas fa-drafting-compass', 'fas fa-lock'],
-  beatmapset_discussion_post_new: ['fas fa-drafting-compass', 'fas fa-comment-medical'],
-  beatmapset_discussion_unlock: ['fas fa-drafting-compass', 'fas fa-unlock'],
-  beatmapset_disqualify: ['fas fa-drafting-compass', 'far fa-times-circle'],
-  beatmapset_love: ['fas fa-drafting-compass', 'fas fa-heart'],
-  beatmapset_nominate: ['fas fa-drafting-compass', 'fas fa-vote-yea'],
-  beatmapset_qualify: ['fas fa-drafting-compass', 'fas fa-check'],
-  beatmapset_reset_nominations: ['fas fa-drafting-compass', 'fas fa-undo'],
-  forum_topic_reply: ['fas fa-comment-medical'],
-  legacy_pm: ['fas fa-envelope'],
-};
 
 export default withMarkRead(observer(class ItemOne extends React.Component<ItemProps & WithMarkReadProps, {}> {
   render() {
@@ -80,13 +64,7 @@ export default withMarkRead(observer(class ItemOne extends React.Component<ItemP
   }
 
   private renderCoverIcon() {
-    const item = this.props.item;
-
-    if (item.name == null || item.category == null) {
-      return null;
-    }
-
-    const icons = ITEM_NAME_ICONS[item.name];
+    const icons = nameToIcons[this.props.item.name || ''];
 
     if (icons == null) {
       return null;
@@ -125,25 +103,9 @@ export default withMarkRead(observer(class ItemOne extends React.Component<ItemP
   }
 
   private renderMessage() {
-    const item = this.props.item;
-    let message: string;
-
-    const replacements = {
-      title: item.details.title,
-      username: item.details.username,
-    };
-
-    const key = `notifications.item.${item.objectType}.${item.category}.${item.name}`;
-
-    if (item instanceof LegacyPmNotification) {
-      message = osu.transChoice(key, item.details.count, replacements);
-    } else {
-      message = osu.trans(key, replacements);
-    }
-
     return (
-      <a href={url(this.props.item)} className='notification-popup-item__row notification-popup-item__row--message clickable-row-link'>
-        {message}
+      <a href={urlOne(this.props.item)} className='notification-popup-item__row notification-popup-item__row--message clickable-row-link'>
+        {messageOne(this.props.item)}
       </a>
     );
   }
